refactoring: change createExecuteToolsTransformation to executeToolsFromStream
