--- packages/ai/src/generate-text/create-execute-tools-transformation.ts
@@ -1,218 +1,0 @@
-import type {
-  Arrayable,
-  Context,
-  IdGenerator,
-  InferToolSetContext,
-  ModelMessage,
-  Sandbox,
-  ToolSet,
-} from '@ai-sdk/provider-utils';
-import type { TimeoutConfiguration } from '../prompt/request-options';
-import type { Telemetry } from '../telemetry/telemetry';
-import { executeToolCall } from './execute-tool-call';
-import { resolveToolApproval } from './resolve-tool-approval';
-import type { LanguageModelStreamPart } from './stream-language-model-call';
-import type { ToolApprovalConfiguration } from './tool-approval-configuration';
-import type { TypedToolCall } from './tool-call';
-import type {
-  OnToolExecutionEndCallback,
-  OnToolExecutionStartCallback,
-} from './tool-execution-events';
-
-export type ToolExecutionEndStreamPart = {
-  type: 'tool-execution-end';
-  toolCallId: string;
-  toolExecutionMs: number;
-};
-
-export type ExecuteToolsStreamPart<TOOLS extends ToolSet = ToolSet> =
-  | LanguageModelStreamPart<TOOLS>
-  | ToolExecutionEndStreamPart;
-
-export function createExecuteToolsTransformation<
-  TOOLS extends ToolSet,
-  RUNTIME_CONTEXT extends Context | unknown | never,
->({
-  tools,
-  callId,
-  messages,
-  abortSignal,
-  timeout,
-  sandbox,
-  toolsContext,
-  toolApproval,
-  runtimeContext,
-  generateId,
-  onToolExecutionStart,
-  onToolExecutionEnd,
-  executeToolInTelemetryContext,
-}: {
-  tools: TOOLS | undefined;
-  callId: string;
-  messages: ModelMessage[];
-  abortSignal: AbortSignal | undefined;
-  timeout?: TimeoutConfiguration<TOOLS>;
-  sandbox?: Sandbox;
-  toolsContext: InferToolSetContext<TOOLS>;
-  toolApproval?: ToolApprovalConfiguration<TOOLS, RUNTIME_CONTEXT>;
-  runtimeContext: RUNTIME_CONTEXT;
-  generateId: IdGenerator;
-  onToolExecutionStart?: Arrayable<OnToolExecutionStartCallback<TOOLS>>;
-  onToolExecutionEnd?: Arrayable<OnToolExecutionEndCallback<TOOLS>>;
-  executeToolInTelemetryContext?: Telemetry['executeTool'];
-}): TransformStream<
-  LanguageModelStreamPart<TOOLS>,
-  ExecuteToolsStreamPart<TOOLS>
-> {
-  const toolCallsToExecute: Array<TypedToolCall<TOOLS>> = [];
-
-  // forward stream
-  return new TransformStream<
-    LanguageModelStreamPart<TOOLS>,
-    ExecuteToolsStreamPart<TOOLS>
-  >({
-    async transform(
-      chunk: LanguageModelStreamPart<TOOLS>,
-      controller: TransformStreamDefaultController<
-        ExecuteToolsStreamPart<TOOLS>
-      >,
-    ) {
-      // immediately forward all chunks
-      controller.enqueue(chunk);
-
-      const chunkType = chunk.type;
-
-      switch (chunkType) {
-        case 'tool-call': {
-          if (chunk.invalid) {
-            return;
-          }
-
-          const tool = tools?.[chunk.toolName];
-
-          if (tool == null) {
-            // ignore tool calls for tools that are not available,
-            // e.g. provider-executed dynamic tools
-            return;
-          }
-
-          const toolApprovalStatus = await resolveToolApproval({
-            tools,
-            toolCall: chunk,
-            toolApproval,
-            messages,
-            toolsContext,
-            runtimeContext,
-          });
-
-          switch (toolApprovalStatus.type) {
-            case 'user-approval': {
-              controller.enqueue({
-                type: 'tool-approval-request',
-                approvalId: generateId(),
-                toolCall: chunk,
-              });
-
-              return; // don't execute tool
-            }
-
-            case 'denied': {
-              const approvalId = generateId();
-
-              controller.enqueue({
-                type: 'tool-approval-request',
-                approvalId,
-                toolCall: chunk,
-                isAutomatic: true,
-              });
-              controller.enqueue({
-                type: 'tool-approval-response',
-                approvalId,
-                approved: false,
-                toolCall: chunk,
-                reason: toolApprovalStatus.reason,
-                providerExecuted: chunk.providerExecuted,
-              });
-
-              return; // don't execute tool
-            }
-
-            case 'approved': {
-              const approvalId = generateId();
-
-              controller.enqueue({
-                type: 'tool-approval-request',
-                approvalId,
-                toolCall: chunk,
-                isAutomatic: true,
-              });
-              controller.enqueue({
-                type: 'tool-approval-response',
-                approvalId,
-                approved: true,
-                toolCall: chunk,
-                reason: toolApprovalStatus.reason,
-                providerExecuted: chunk.providerExecuted,
-              });
-
-              break; // continue with tool execution
-            }
-
-            case 'not-applicable':
-              break; // continue with tool execution
-          }
-
-          // Only execute tools that are not provider-executed:
-          if (tool.execute != null && chunk.providerExecuted !== true) {
-            toolCallsToExecute.push(chunk);
-          }
-
-          return;
-        }
-
-        case 'model-call-end': {
-          await Promise.all(
-            toolCallsToExecute.map(async toolCall => {
-              try {
-                // Note: we don't await the tool execution here (by leaving out 'await' on recordSpan),
-                // because we want to process the next chunk as soon as possible.
-                // This is important for the case where the tool execution takes a long time.
-                const result = await executeToolCall({
-                  toolCall,
-                  tools,
-                  callId,
-                  messages,
-                  abortSignal,
-                  timeout,
-                  sandbox,
-                  toolsContext,
-                  onToolExecutionStart,
-                  onToolExecutionEnd,
-                  executeToolInTelemetryContext,
-                  onPreliminaryToolResult: result => {
-                    controller.enqueue(result);
-                  },
-                });
-                if (result != null) {
-                  controller.enqueue({
-                    type: 'tool-execution-end',
-                    toolCallId: result.output.toolCallId,
-                    toolExecutionMs: result.toolExecutionMs,
-                  });
-                  controller.enqueue(result.output);
-                }
-              } catch (error) {
-                controller.enqueue({
-                  type: 'error',
-                  error,
-                });
-              }
-            }),
-          );
-
-          return;
-        }
-      }
-    },
-  });
-}
